pyproject.toml
[project]
name = "travel-assistant-rag"
version = "1.0.0"
description = "Enterprise RAG-Enhanced Travel Assistant with LangGraph + Qdrant + Gemini"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Vector DB & Search
    "qdrant-client>=1.7.0",
    
    # Google Gemini
    "google-generativeai>=0.3.0",
    "langchain-google-genai>=0.0.5",
    
    # LangChain/LangGraph
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.20",
    
    # API Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # UI
    "gradio>=4.15.0",
    
    # Utilities
    "httpx>=0.26.0",
    "tenacity>=8.2.0",
    "rich>=13.7.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]
